Put changes to preprocess into original file
